requirements.txt
accelerate==1.6.0
addict==2.4.0
aiofiles==23.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.12
aiosignal==1.3.2
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.3.0
blinker==1.9.0
ccimport==0.4.4
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
coloredlogs==15.0.1
colorlog==6.9.0
comm==0.2.2
ConfigArgParse==1.7.1
contourpy==1.3.1
cumm-cu126==0.7.11
cycler==0.12.1
dash==3.3.0
debugpy==1.8.14
decorator==5.2.1
docstring_parser==0.16
einops==0.8.1
embreex==2.17.7.post6
exceptiongroup==1.2.2
executing==2.2.0
fastapi==0.115.12
fastjsonschema==2.21.2
ffmpy==0.5.0
filelock==3.18.0
fire==0.7.0
Flask==3.1.2
flatbuffers==25.2.10
fonttools==4.57.0
frozenlist==1.7.0
fsspec==2025.3.2
glcontext==3.0.0
gradio==5.17.1
gradio_client==1.7.1
h11==0.14.0
h5py==3.13.0
httpcore==1.0.8
httpx==0.28.1
huggingface-hub==0.30.2
humanfriendly==10.0
hydra-core==1.3.2
idna==3.10
imageio==2.37.0
importlib_metadata==8.7.0
ipdb==0.13.13
ipykernel==6.29.5
ipython==8.35.0
ipywidgets==8.1.8
itsdangerous==2.2.0
jaxtyping==0.3.2
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_widgets==3.0.16
kiwisolver==1.4.8
lark==1.2.2
lazy_loader==0.4
lightning-utilities==0.14.3
lpips==0.1.4
lxml==5.3.2
manifold3d==3.0.1
mapbox_earcut==1.0.3
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.10.1
matplotlib-inline==0.1.7
mdurl==0.1.2
moderngl==5.12.0
mpmath==1.3.0
msgspec==0.19.0
multidict==6.4.4
munch==4.0.0
narwhals==2.12.0
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
ninja==1.11.1.4
nodeenv==1.9.1
numpy==1.26.1
numpy-quaternion==2024.0.12
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
omegaconf==2.3.0
onnxruntime==1.21.0
open3d==0.19.0
opencv-python==4.11.0.86
orjson==3.10.16
packaging==24.2
pandas==2.2.3
parso==0.8.4
pccm==0.4.16
pexpect==4.9.0
pillow==11.2.1
platformdirs==4.3.7
plotly==6.5.0
plyfile==1.1
portalocker==3.2.0
prompt_toolkit==3.0.50
propcache==0.3.2
protobuf==6.30.2
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pybind11==2.13.6
pycollada==0.9
pydantic==2.10.6
pydantic_core==2.27.2
pydub==0.25.1
Pygments==2.19.1
pyparsing==3.2.3
pyquaternion==0.9.9
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytorch-lightning==2.5.1.post0
pytz==2025.2
PyYAML==6.0.2
pyzmq==26.4.0
referencing==0.36.2
requests==2.32.3
retrying==1.4.2
rich==14.0.0
rpds-py==0.24.0
rtree==1.4.0
ruff==0.11.5
safehttpx==0.1.6
safetensors==0.5.3
scikit-image==0.25.2
scikit-learn==1.7.2
scipy==1.15.2
semantic-version==2.10.0
shapely==2.1.0
shellingham==1.5.4
shtab==1.7.2
six==1.17.0
sniffio==1.3.1
spconv-cu126==2.3.8
stack-data==0.6.3
starlette==0.46.2
svg.path==6.3
sympy==1.13.1
taming-transformers==0.0.1
termcolor==3.1.0
threadpoolctl==3.6.0
tifffile==2025.3.30
tomli==2.2.1
tomlkit==0.13.2
torch==2.6.0+cu124
torch_cluster==1.6.3+pt26cu124
torch_scatter==2.1.2+pt26cu124
torchmetrics==1.7.2
torchvision=0.21.0+cu124
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
trimesh==4.6.6
triton==3.2.0
typeguard==4.4.2
typer==0.15.2
typing-inspection==0.4.0
typing_extensions==4.13.2
tyro==0.9.18
tzdata==2025.2
urllib3==2.4.0
utils3d @ git+https://github.com/EasternJournalist/utils3d.git@c5daf6f6c244d251f252102d09e9b7bcef791a38
uvicorn==0.34.1
vhacdx==0.0.8.post2
viser==0.2.23
wadler_lindig==0.1.6
wcwidth==0.2.13
websockets==14.2
Werkzeug==3.1.3
widgetsnbextension==4.0.15
xxhash==3.5.0
yarl==1.20.1
yourdfpy==0.0.57
zipp==3.23.0
